sprint_2-week_4-homeTask_2 - added some minor fix

+++ src/users/repositories/user.repository.ts
@@ -19,6 +19,9 @@ export class UserRepository {
   }
 
   async deleteUserRepo(id: string): Promise<boolean> {
+    // * Проверяем, является ли ObjectId действительным
+    if (!ObjectId.isValid(id)) return false;
+
     const isDeleted = await userCollection.deleteOne({ _id: new ObjectId(id) });
 
     return isDeleted.deletedCount === 1;
